pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ceramicraft-log-mservice"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dttb>=0.3.2",
    "grpcio>=1.78.0",
    "protobuf>=7.34.0",
    "psycopg[binary]>=3.3.3",
    "python-dotenv>=1.2.2",
    "sqlalchemy>=2.0.48",
    "typer>=0.24.1",
]

[dependency-groups]
test = [
    "pytest>=9.0.2",
    "testcontainers[postgres]>=4.14.1",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
